requirements.txt
dspy-ai==3.0.4
python-dotenv==1.2.1
